workflow will load existing values fixes #73

+++ Analysis/main.R
@@ -5,60 +5,91 @@ source('prepareData.R')
 
 if (length(experiments)==length(marketFiles)) {
   dir.create("output data")
+  
   for (i in 1:length(experiments)) {
     experimentName <- experiments[i]
+    marketFile <- marketFiles[i]
     path <- paste("output data/", experimentName, "/", sep="")
-    #if the folder already exists delete all the content 
-    if (file.exists(path)) {
+    
+    # if experiment folder does not exsits create it 
+    if (!file.exists(path)){
+      dir.create(path)
+    }
+    
+    # check for the experiment configuration 
+    pathConfig <- paste(path, "experimentConfig.tsv", sep="")
+    if (file.exists(pathConfig)) {
+      cpdf <- read.table(pathConfig, header=TRUE, sep="\t", stringsAsFactors=FALSE)
+      print(path)
+      print(pathConfig)
+      if (cpdf[cpdf$parameter=="file",]$value!=marketFile | cpdf[cpdf$parameter=="start year",]$value!=start |
+          cpdf[cpdf$parameter=="end year",]$value!=end | cpdf[cpdf$parameter=="moving average used",]$value!=useMovingAverage |
+          cpdf[cpdf$parameter=="multiplication factor",]$value!=multiplicationFactor) {
+        # delete everything that exists 
+        unlink(path, recursive = TRUE)
+        dir.create(path)
+        # save config parameters from the experiment 
+        cpdf <- data.frame(parameter=c("experiment name", "file", "start year", "end year", "moving average used", "prediction years", "multiplication factor"), 
+                           value=c(experimentName, marketFile, start, end, useMovingAverage, paste(predictionYears, collapse=","), multiplicationFactor))
+        write.table(cpdf, file=pathConfig, quote=FALSE, sep="\t", 
+                    col.names=TRUE, row.names=FALSE)
+      }
+    } else {
+      # if experimentConfig.tsv does not exsist delete everything else
       unlink(path, recursive = TRUE)
     }
-    # create expeirment folder 
-    dir.create(path)
-    
-    marketFile <- marketFiles[i]
-
-    # save the config parameters from the experiment 
-    cpdf <- data.frame(parameter=c("experiment name", "file", "start year", "end year", "moving average used", "prediction years", "multiplication factor"), 
-                       value=c(experimentName, marketFile, start, end, useMovingAverage, paste(predictionYears, collapse=","), multiplicationFactor))
-    write.table(cpdf, file=paste(path, "experimentConfig.tsv", sep=""), quote=FALSE, sep="\t", 
-                col.names=FALSE, row.names=FALSE)
-    
+    rm(pathConfig)
     
     # load market data 
     # load it twice to see which columns it has
     marketData <- read.table(paste("input data/", marketFile, sep=""), header=TRUE, sep="\t", stringsAsFactors=FALSE)
     marketData <- read.table(paste("input data/", marketFile, sep=""), header=TRUE, 
-                            colClasses=rep("character", each=ncol(marketData)), sep="\t", stringsAsFactors=FALSE)
+                             colClasses=rep("character", each=ncol(marketData)), sep="\t", stringsAsFactors=FALSE)
     
-
     # calculate properties to take TODO explain exclusion from the tsv file 
     nam <- names(marketData)
-    propertyToTake <- nam[!(nam %in% c("ID", "name", "release.year", "comments"))]  
+    propertyToTake <- nam[!(nam %in% c("ID", "name", "release.year", "comments"))]
     
-    # discard those elements from the market where properties, according to which filtering should
-    # be done, are missing
-    source('discardNotComplete.R')
-    marketData <- discardNotComplete(marketData, propertyToTake)
-    write.table(marketData, file=paste(path, "market.tsv", sep=""), quote=FALSE, sep="\t", 
-                col.names=TRUE, row.names=FALSE)
-    
-    # make a list from multiple mime and other property values
-    for (prop in propertyToTake) {
-      marketData[[prop]] <- strsplit(marketData[[prop]], split = ",")
+    # check if dataCleaned.tsv exsists and if yes load it 
+    # otherwise prepare data 
+    dataCleaned <- NA
+    pathDataCleaned <- paste(path,"cleanedData.tsv",sep="")
+    if (file.exists(pathDataCleaned)) {
+      dataCleaned <- read.table(pathDataCleaned, header=TRUE, sep="\t", stringsAsFactors=FALSE)
+      cat("Cleaned data loaded from the disk")
+    } else {
+      # discard those elements from the market where properties, according to which filtering should
+      # be done, are missing
+      source('discardNotComplete.R')
+      marketData <- discardNotComplete(marketData, propertyToTake)
+      write.table(marketData, file=paste(path, "market.tsv", sep=""), quote=FALSE, sep="\t", 
+                  col.names=TRUE, row.names=FALSE)
+      
+      # make a list from multiple mime and other property values
+      for (prop in propertyToTake) {
+        marketData[[prop]] <- strsplit(marketData[[prop]], split = ",")
+      }
+      
+      # load raw data, filter according to properties, reduce conflicts calculate age and save the resulting dataset to a file  
+      dataCleaned <- prepareData(fileName, colNames, marketData, propertyToTake, resolveConflicts, 
+                                 afterConflictsResolution, conflictCategory)
+      write.table(dataCleaned, file=paste(path,"cleanedData.tsv",sep=""), quote=FALSE, 
+                  sep="\t", col.names=TRUE, row.names=FALSE)
     }
+    rm(pathDataCleaned)
     
-    # load raw data, filter according to properties, reduce conflicts calculate age and save the resulting dataset to a file  
-    dataCleaned <- prepareData(fileName, colNames, marketData, propertyToTake, resolveConflicts, 
-                     afterConflictsResolution, conflictCategory)
-    write.table(dataCleaned, file=paste(path,"cleanedData.tsv",sep=""), quote=FALSE, 
-                sep="\t", col.names=TRUE, row.names=FALSE)
-    
-    # calculate percentage and moving average of each value
-    source('normalizeMarket.R')
-    dataShares <- normalizeMarket(dataCleaned,propertyToTake, multiplicationFactor)
-    write.table(dataShares, file=paste(path, "adoptionRates.tsv", sep=""), 
-                quote=FALSE, sep="\t", col.names=TRUE, row.names=FALSE)
-    
+    pathDataShares <- paste(path, "adoptionRatse.tsv", sep="")
+    dataShares <- NA
+    if (file.exists(pathDataShares)) {
+      dataShares <- read.table(pathDataShares, header=TRUE, sep="\t", stringsAsFactors=FALSE) 
+    } else {
+      # calculate percentage and moving average of each value
+      source('normalizeMarket.R')
+      dataShares <- normalizeMarket(dataCleaned,propertyToTake, multiplicationFactor)
+      write.table(dataShares, file=paste(path, "adoptionRates.tsv", sep=""), 
+                  quote=FALSE, sep="\t", col.names=TRUE, row.names=FALSE)
+    }
+    rm(pathDataShares)
     
     # estimating models, plotting results and making predictions 
     source('estimateModelParameters.R')
@@ -66,15 +97,42 @@ if (length(experiments)==length(marketFiles)) {
     
     # create folder where all models will be saved 
     pathMarketElements <- paste(path, "market elements/", sep="")
-    dir.create(pathMarketElements)
+    if (!file.exists(pathMarketElements)){
+      dir.create(pathMarketElements)
+    } 
     
-    # for each element estimate models 
-    allModelEstimates <- estimateModelParameters(dataShares, start, end, useMovingAverage, multiplicationFactor, pathMarketElements)
-    saveRDS(allModelEstimates, file=paste(pathMarketElements, "allModelEstimates.rds", sep=""))
-    bestModelEstimates <- calculateBestModelValues(dataShares, allModelEstimates, pathMarketElements)
-    saveRDS(bestModelEstimates, file=paste(pathMarketElements, "bestModelEstimates.rds", sep="")) 
-    # plot bets models
+    pathAllModelEstimates <- paste(pathMarketElements, "allModelEstimates.rds", sep="")
+    allModelEstimates <- NA
+    if (file.exists(pathAllModelEstimates)){
+      allModelEstimates <- readRDS(pathAllModelEstimates)
+    } else {
+      # for each element estimate models 
+      allModelEstimates <- estimateModelParameters(dataShares, start, end, useMovingAverage, multiplicationFactor, pathMarketElements)
+      saveRDS(allModelEstimates, file=paste(pathMarketElements, "allModelEstimates.rds", sep=""))
+    }
+    rm(pathAllModelEstimates)
+    
+    pathBestModelEstimates <- paste(pathMarketElements, "bestModelEstimates.rds", sep="")
+    bestModelEstimates <- NA
+    if (file.exists(pathBestModelEstimates)) {
+      bestModelEstimates <- readRDS(pathBestModelEstimates)
+    } else {
+      bestModelEstimates <- calculateBestModelValues(dataShares, allModelEstimates, pathMarketElements)
+      saveRDS(bestModelEstimates, file=paste(pathMarketElements, "bestModelEstimates.rds", sep="")) 
+    }
+    
+    # plot best models
     graphAllModels(dataShares, bestModelEstimates, pathMarketElements)
+    
+    # sync graphs to the selected folder 
+    for (name in unique(dataShares$name) ) {
+      pathFrom <- paste(pathMarketElements, name, "/graphs", sep="")
+      print(pathFrom)
+      file.copy(from=list.files(pathFrom, full.names = TRUE) , to=syncFolder, 
+                overwrite = TRUE, recursive = TRUE, 
+                copy.mode = TRUE)
+    }
+    
   }
   
   #Ask user to select wanted models manually 
@@ -101,30 +159,30 @@ if (length(experiments)==length(marketFiles)) {
     
   }
   
-  syncFolder <- "/Users/kresimir/Dropbox/Work/Projects/BenchmarkDP/formatanalysis/figures/201605 figures/20160506-1200/model figures/"
-  for (i in 1:length(experiments)) {
-    experimentName <- experiments[i]
-    path <- paste("output data/", experimentName, "/", sep="")
-    print(path)
-    dataShares <- read.table(paste(path, "adoptionRates.tsv", sep=""), header=TRUE, sep="\t", stringsAsFactors=FALSE)
-    pathElements <- paste("output data/", experimentName, "/market elements", sep="")
-    
-    for (name in unique(dataShares$name) ) {
-      pathFrom <- paste(pathElements, "/", name, "/graphs", sep="")
-      print(pathFrom)
-      file.copy(from=list.files(pathFrom, full.names = TRUE) , to=syncFolder, 
-                overwrite = TRUE, recursive = TRUE, 
-                copy.mode = TRUE)
-    }
-    
-    pathAllGraphs <- paste(path, "graphs", sep="")
-    print(pathAllGraphs)
-    file.copy(from=list.files(pathAllGraphs, full.names = TRUE) , to=syncFolder, 
-              overwrite = TRUE, recursive = TRUE, 
-              copy.mode = TRUE)
-    
-  }
-  
+#   syncFolder <- "/Users/kresimir/Dropbox/Work/Projects/BenchmarkDP/formatanalysis/figures/201605 figures/20160506-1200/model figures/"
+#   for (i in 1:length(experiments)) {
+#     experimentName <- experiments[i]
+#     path <- paste("output data/", experimentName, "/", sep="")
+#     print(path)
+#     dataShares <- read.table(paste(path, "adoptionRates.tsv", sep=""), header=TRUE, sep="\t", stringsAsFactors=FALSE)
+#     pathElements <- paste("output data/", experimentName, "/market elements", sep="")
+#     
+#     for (name in unique(dataShares$name) ) {
+#       pathFrom <- paste(pathElements, "/", name, "/graphs", sep="")
+#       print(pathFrom)
+#       file.copy(from=list.files(pathFrom, full.names = TRUE) , to=syncFolder, 
+#                 overwrite = TRUE, recursive = TRUE, 
+#                 copy.mode = TRUE)
+#     }
+#     
+#     pathAllGraphs <- paste(path, "graphs", sep="")
+#     print(pathAllGraphs)
+#     file.copy(from=list.files(pathAllGraphs, full.names = TRUE) , to=syncFolder, 
+#               overwrite = TRUE, recursive = TRUE, 
+#               copy.mode = TRUE)
+#     
+#   }
+#   
   
 } else {
   message("The number of experiments and market files does not match! Check experiments and marketFiles variables in the config.R file")
